+++ packages/toolkit/project/local/scripts/build.sh
@@ -1,0 +1,5 @@
+# Add builds scripts here. This script will be run from the root of your project (same directory as .tenup.yml).
+
+nvm install
+node -v
+npm install